refactor some kits related function
